Version 1.1 ~10.9 eggs/hour
Works with any egg cycle length (20 or under).  Character will now hatch the egg as soon as it is ready and then continue to walk for the remainder of the 20 egg cycles.  This vastly decreases the chances of the Nursery Lady not having an egg ready, especially for Pokemon with short egg cycles.  Feel free to comment out egg cycles if you would like to optimize this for a specific pokemon.